Revert commit r2258 which made texture coordinates into floating point. It caused bug #9631, wrong team colours displayed.
git-svn-id: svn+ssh://svn.gna.org/svn/warzone/trunk@2292 4a71c877-e1ca-e34f-864e-861f7616d084

+++ lib/ivis_common/bitimage.h
@@ -35,14 +35,14 @@ static inline WZ_DECL_PURE unsigned short iV_GetImageHeight(const IMAGEFILE *Ima
 }
 
 
-static inline WZ_DECL_PURE float iV_GetImageXOffset(const IMAGEFILE *ImageFile, const unsigned short ID)
+static inline WZ_DECL_PURE short iV_GetImageXOffset(const IMAGEFILE *ImageFile, const unsigned short ID)
 {
 	assert(ID < ImageFile->Header.NumImages);
 	return ImageFile->ImageDefs[ID].XOffset;
 }
 
 
-static inline WZ_DECL_PURE float iV_GetImageYOffset(const IMAGEFILE *ImageFile, const unsigned short ID)
+static inline WZ_DECL_PURE short iV_GetImageYOffset(const IMAGEFILE *ImageFile, const unsigned short ID)
 {
 	assert(ID < ImageFile->Header.NumImages);
 	return ImageFile->ImageDefs[ID].YOffset;
